+++ src/Products/MeetingNamur/adapters.py
@@ -46,6 +46,8 @@ customWfAdaptations = (
     'item_validation_shortcuts',
     'item_validation_no_validate_shortcuts',
     'only_creator_may_delete',
+    'meeting_remove_global_access',
+    'meetingmanager_correct_closed_meeting',
     # first define meeting workflow state removal
     'no_freeze',
     'no_publication',
@@ -71,7 +73,6 @@ customWfAdaptations = (
     'accepted_out_of_meeting_emergency_and_duplicated',
     'transfered',
     'transfered_and_duplicated',
-    'meetingmanager_correct_closed_meeting',
     'namur_meetingmanager_may_not_edit_decision_project',
 )
 MeetingConfig.wfAdaptations = customWfAdaptations